Add support for running llvm sanitizers on Android
On Android, ASan requires to push the libclang_rt.so
shared library on the device. This CL reorganizes the
build files and adds a target to copy the library in
the output directory. Also adds a convenience script
to run tests on the device that takes care of pushing
the binary and the required libs.

Change-Id: I9710a2019c91748934b58951235aec025bb3d96a