[IPC 5] Add host side logic
Introduces the host part. HostImpl is the class that
owns the listening socket, parses incoming requests
and dispatches them to the service instances.
The lifetime of services and the Host is decoupled
via weak pointers.

The final picture for the host is as follows:

                           / ServiceA
UnixSocket <- Host(Impl) <-- ServiceB
                           \ ServiceC

End-to-end tests and codegen plugin will follow in
next CLs.

Bug: 68854111
Change-Id: I59c835639b784b5d6480a04a1acab76e82b78141