Generate full protos
For testing we would like to be able to use protos reflection APIs
(comparison, to/from string, etc) unfortunately this is precisely what is
not present when using the lite runtime which is what we would like to use
in production. Even worse for reasons unknown the choice of runtime is
specified *in* the proto file rather than at generation or compile time.

To get around this we generate a copy of all the proto files without the
runtime specification, then use this to generate non-lite headers.

Change-Id: I9ad69ebcb195cddd9487378d2dd8754ad497b345