Move WeakPtrFactory as last field: improves tests robustness.
As the destruction order goes bottom-up (w.r.t. declaration order of
fields), having the WeakPtrFactory as last field makes it so that
the weak ptrs are invalidated as first thing during destruction.
This makes it easier to write tests without having to worry too much
about destruction order.

Also this CL fixes a test flakiness due to the fact that
TestSharedMemory didn't return page-aligned memory.

Change-Id: If7218e04a138b4e69170594bf5dec723b5afe7bb