Fix trace_to_text host binary generation
This CL fixed a bug in the proto_library.gni, that rebases
the proto file onto the |root_out_dir| rather than |root_build_dir|.
This causes an extra ../ in the invocation of protoc.
This also makes it so that the trace_to_text is built only for the
host but the binary is copied in the root build directory.

Change-Id: Ib50e6a5dd949d197fc3d5780aad2e9574ec34111