Remove usage of linkat()
Remove usage of linkat() in favor of making a temporary file in the
filesystem while tracing. The SELinux policy doesn't allow for the use
of hard links.

A future patch will migrate the command line client to a new DropBox API
which directly accepts a file descriptor, removing the need to create a
filesystem-visible file.

Change-Id: I2b5480387f1c459fc885234245fe4403ffd5190e