Implement Track Registry
Implements a track registry so the track implementation can be loaded
dynamically. Adds barebones implementation of two track types
(CpuSliceTrack and CpuCounterTrack) for demoing the registry.

The track implementation is currently represented by TrackImpl class,
which can only draw on canvas using the draw() method. How a track will
implement DOM is still TBD and out of the scope for this CL.


Bug:111540363

Change-Id: I0489f161c3783af6d54568fe839301f47ae0291a