TraceProcessor: fix cost estimation of process and thread vtables
The current implementation was mistakenly returning a cost == 0
for queries without any constraint, tricking the SQLite query
engine into believing that a full table scan without any constraints
was cheaper then a query with a filter on utid/upid.
This fixes the cost estimation, returning by default the number
of rows in the table scan.

In practice, the time it takes to execute the following query:
select count(*) from sched \
  inner join thread using(utid) inner join process using(upid)

Goes down from ~30s to ~5s for a 30 seconds trace.

This CL also fixes a bug in the parsing logic that was causing
the trace processor to hang forever in presence of a truncated
trace.

Change-Id: I09cb667eb6f05a3302217490cb535d847706d596