Exclude watchdog_posix.cc in AOSP MAC builds
This CL #ifdefs out the watchdog code on AOSP mac builds.
Recently we started pulling this target also for host executables
but watchdog_posix.cc doesn't support Mac. This is not a problem
for GN-based builds, where we exclude the file on mac. In AOSP,
however, we don't propagate the Mac exclusion in the build system,
because that is too hard and not worth the effort.
Instead we just ifdef out the file on mac.

Change-Id: Ia85ed182f155ba5cb36aaba956c3e7ea60ed0800