trace_processor: change cursor class to be scoped to a filter operation
Before this point, our cursors were structured in the same way as SQLite
cursors. That is, a cursor object can be reused for multiple filter
operations.

However, in pretty much all our cursor cases, the state of the cursor is
tied to a filter operation and not the scope of the SQLite cursor.

Change the scoping by creating a RawCursor class at the Table level
which handles conversion of scoping between SQLite filters and creation
of our cursors.

Bonus: this clears up a lot of "FilterState" and similar classes which
were scoping to a filter operation in the all the sub classes.

This also allows for neater implementation of multiple algorithms for
traversing a table (e.g. upcoming multi algorithm span tables).

Change-Id: I5068a9868c2271b6ab4655f61fcee5659b17c328