trace_processor: fix sorting of events in sched tracker
Problem:
Sched and counter events arrive in order at the sched tracker but exit
out of order when placed in the trace storage. We make assumptions that
the data in trace storage is always sorted and this assumption does not
hold if sched tracker does not output events as it receives them.

Reason:
Computation of duration and other parameters which require both the begin
and end events mean that until now, we were waiting for the end event
before emitting the slice/counter. However, ends may happen out of order
with respect to the start timestamp of the slice.

Solution:
Place the data in trace storage as soon as it received and backfill data
which cannot be computed straight away. Also switch from prev_comm to
next_comm to reduce backfill data for sched events.

Change-Id: Icecb4e409b6904a6f0d6dd046e9fbf0028968aa8