perfetto-ui: Fix jank when mouse leaves tracks
When panning with the mouse we happen to schedule a repaint when
the mouse is moving over the tracks but not when scrolling over
non-track content. We do a full redraw (with dom) every ~100Hz
so this manifests to as terrible jank when the mouse leaves the track
area.

This changes ensures we repaint whenever we scroll.

Change-Id: Idd5cefb02e87d421111b23ecccb07bda671d12c0