trace_processor: fix handling of hidden columns
During the schema refactor, we ended up in a state where hidden columns
weren't actually reported as hidden to SQLite. This caused the hidden
columns in the window table to be visible in SPAN JOINs when they
shouldn't be so.

Readd handling for this and add tests to check output of window table to
make sure this doesn't break.

Change-Id: If8d0ec74ddff814763d7c7a6b4ed3f899397471b