trace_processor: fix ordering of slices in trace storage
Slices are meant to be stored ordered by timestamp. However, because we
were waiting until the end event to push the slice to storage, each
stack would actually be stored in *reverse* timestamp order.

Fix this issue by pushing the slice to storage as soon as we see the
start event and then updating duration after the fact. This works
similarily to sched events.

Change-Id: If89ff8f4fde91917b090f1013efafd0f3d9d997d