trace_processor: fix multiple bugs with filtering and casting
When comparision against doubles was attempted to be fixed in aosp/821500,
we accidentally broke comparision of integers vs integers. Moreover,
by casting doubles to integers, we were losing precision in our
comparisions. This caused 119496959 which led to bugs in the timeline.

Fix these issues by choosing more carefully which type we want to
compare in - compare as integers if both the sqlite value and the value
in the table are integers. Otherwise, convert both to doubles and
compare there.

Also fix a minor bug where sqlite_value_int was used for int64 types.

Bug: 119496959
Change-Id: I6369abed91b9c97f58ec7da3f314d1ebbacdbab4