Add generic ftrace event
Any ftrace event that perfetto does not have prior knowledge of
can be added to the config and outputted in the form of a generic
event.

The fields outputted in the generic proto are populated based on the
fields read from the ftrace format file of that event.

This CL also adds logic for the config to understand ftrace events
that include the group name e.g sched/sched_switch. In order for
a generic event to be used the group must be included. For backward
compatibility, events can still be specified without a group.

Bug:117471443
Change-Id: I96a8565b33822ef3fadf37f9ff9904d574fec6f8