Fix null pointer dereference when using --attach
The RateLimiter code was assuming that the TraceConfig is always
populated in the trace_config_->guardrail_overrides() ... check.
But trace_config_ was nullptr when using --attach, causing a NPE.
Unfortunatly the one line was behind the a
PERFETTO_ANDROID_USERDEBUG_BUILD compile check so bypassed the
coverage in the standalone CI.
This CL fixes the bug by always assigning a TraceConfig instance
and makes it so the standalone build also covers that one line.

Test: manual + more coming in aosp/856349
Bug: 120607375
Change-Id: Ibb583a65271ce13cb6184d7f7652def945bb8481