perfetto-ui: Extract constants and read from css
We have a couple of constants relating to tracks which are duplicated
over the code base (and sometimes in CSS). Extract these into a single
file and (where necessary) read the value from the CSS.

In order to be able to test the codebase without CSS if we are loading
in a non-browser environment we use a default value instead.

Change-Id: I4c5361495d887cd48afb008f5c0ab4d9ea7e21cd