perfetto-ui: Stop frontend DOSing controller
The front end sends a request for more data once as soon as it runs out
and then every 50ms thereafter. With many tracks requesting data at the
same time this can easily DOS the controller. As a temporary hack bump
the 50ms figure for summary tracks to 250ms. The real fix to batch
requests for more data and (ideally) have them driven by the controller.

Bug: 123616604
Change-Id: I0c9399527a70b84be74539812064b7a4a9c9ef31