protozero: emit empty .cc files
Moves everything to the header.
There was very little left in the .cc files:
1) The set_xxx, which should have been inlined since the beginning.
2) The reflection methods (not really used)
Makes the build ~4s. faster on a macbook pro 2018.

We still emit (now empty) .cc files. This is because the
chromium proto_library template expects the generator to
generate both .h and .cc. Will fix that in a separate step.

Bug: 74184542
Change-Id: I7f8c3fbb9703c0e6b6419a6bfc7e6a3866d22892