arbiter: Only flush synchronously on ProducerEndpoint thread
Currently, the arbiter can end up causing CommitRequests to be
reordered if a synchronous flush on a different thread "steals" an old
request. (This would cause a task to be posted to the endpoint thread
that may execute after a previously scheduled flush task. The latter
task may notify the endpoint of later commits if any have occurred in
the time between the synchronous flush and the task executing.) See
crbug.com/919187 for details.

This patch avoids this problem by performing synchronous flushes only on
the same thread that notifies the producer endpoint.

Bug: chromium:919187
Change-Id: I762046abc368f41a787a1b05abf9d08dc0af73d2