heapprofd: fix unwinding of native frames in apk-embedded native libs
Back in October, the signature of MapInfo's constructor changed, and
aosp/778458 added a fixed nullptr argument in heapprofd (presumably as a
placeholder). This hasn't been changed since, and makes libunwindstack
interpret every mapping as the first (i.e. having no predecessor). This missing
predecessor link is important when trying to look up certain cases of mmap'd
apk-embdedded dso's, and is the cause for heapprofd's failures to handle such
cases.

The particular case that wasn't being handled is when the embedded dso's header
& text section are in separate mappings:
7124158000-7124159000 r--p 00049000 fd:06 19024                          /data/app/android.perfetto.cts.app.debuggable-Hxh9ZCB0iPsJX1eM4v-3OA==/base.apk
7124159000-712415a000 --xp 0004a000 fd:06 19024                          /data/app/android.perfetto.cts.app.debuggable-Hxh9ZCB0iPsJX1eM4v-3OA==/base.apk

Unwinding code:
1) resolved an address to the executable mapping (i.e. the segment containing the .text section).
2) verified that it's not a well-formed ELF, and attempted several fallbacks.
3) the fallback that was supposed to work (look at previous mapping, see if
   it's the start of the relevant elf) didn't, since heapprofd didn't populate the
   prev_map linked list correctly.
4) rest of the code didn't handle this situation. And ended up trying to
   use just the executable segment as the elf. The parsing failed, but
   libunwindstack didn't catch the error in that codepath.

With this patch, heapprofd can now unwind the native frames in apk-embedded
dso's (at least for the cases of readable apk files). If fully precompiling
the app, we can now see full callstack (jni_lib->art->java).

*However* we still fail to unwind into art/java when running under interpreter
alone. I'm investigating.

This is only the perfetto side of the fix, I also want to send out a patch to
libunwindstack s.t. failures like this are reported (instead of being silently
broken).

Change-Id: Ie7cb754a30e88bf4a8870f8b45663ee605076796