heapprofd: start task duration watchdog (default of 30s)
Tested manually. Automated testing would require a small refactoring that I
don't want to be part of this cl.

Note that the watchdog will start a thread that's useless for heapprofd (since
we're not doing cpu/mem limit tracking). WDYT, worth adding a TODO to add a
lighter-weight watchdog setup?

Change-Id: I1a17ff0149d7a30d1fd5300da3debd979aed0d53