trace_processor: print non-zero error stats and stop counter events from log spam
While the stats table is very important for putting data that we consider
important context to trace analysis, often people can overlook looking at
it until after some significant investigation has been done. This can lead
to frustration that these stats were not surfaced earlier.

Change this by querying stats at the end of importing the trace in trace
processor shell and printing it out to stderr. This allows users to see if their
trace is busted/trace processor has a bug before they get into proper analysis.

And since we have stats for this now and print them in shell,
turn the ELOG into a DLOG for counter out of order
to avoid log spam if this does happen.

Bug: 126443808
Change-Id: I2cdb7e2710242d044c78f729c004ab93c38fcc90