SMA: Flush pending commits more frequently during SMB stalls
Previously, we were flushing commits only when the stall count in
GetNewChunk() reached exactly 3. This is problematic, particularly in
in-process tracing mode:

If the service is in-process on the same thread/task sequence which
stalls, it is possible that other threads are writing more trace data
concurrently. Previously, the stalling service sequence would
eventually flush the pending commits in GetNewChunk(), but then enter
a <=100ms sleep. During this sleep, other threads can write more data
into the SMB again, even filling it up again. (This is theoretically
also possible without the sleep, but maybe less likely.) Once the
service thread wakes up from its sleep, it will attempt to obtain a new
chunk again and continue to stall (this time forever).

To resolve this potential issue, the service/IPC thread should
attempt to flush pending commits more often - and also avoid the sleep
after a flush.

More precisely, the problem is likely to occur in this scenario with
an in-process perfetto service (IPC thread == service thread):
1. Another thread A fills up the SMB with N chunks and notifies the
   IPC/service thread to flush soon.
2. The IPC/service thread itself attempts to trace and is stalled waiting
   for a new chunk. It synchronously flushes the pending commits.
3. The IPC/service thread synchronoulsy starts committing them into the
   central buffer, freeing up all of the SMB pages and then is
   descheduled.
4. Concurrently, thread A fills up and commits N chunks again before the
   IPC/service thread gets to run again. The SMB is full again.

In this scenario the SMB is full, the IPC/service thread is still
stalling, but it will not tell the service to commit those new N chunks
ever again.

Change-Id: I6699ab39d56a109865c4edc68f7c3de433ada82b