Make test trace writer keep its buffer as proper trace packet protos
Before this change, the test trace writer stored only the contents of
the trace packets (without the tag & length for the enclosing packet
itself). This corresponds to the real trace writer's behaviour, but
complicates the inspection of trace contents within the tests.

The previous implementation only really made sense for at most a single
packet, as otherwise it would merge everything into one packet
(following the specifics of proto merging rules).

Bug: 132394650
Change-Id: If9077c5733addcce6709088ff25c5beac6faf98c