probes: invalidate seen pids on task_rename event
Task rename events on Android often signal also a change in the name
of the process (think zygote specializing to become an app - this
changes both the name of the main thread and the name of the process).
Since we don't rescrape a pid once we've ever seen it (modulo
incremental state invalidation), we never see the updated names for
these processes either.

Rescrape any pid which undergoes a task rename by removing it from the
seen pids. This is a heuristic (as processes many rename without thread
renames and threads which are not the main process often get renamed)
but it is a good one as it covers the cases we are most interested in.

Bug: 131733942
Bug: 130543265
Change-Id: If20c3518391f8b766b8080c23886cbc145fcae59