perfetto-ui: Fix performance issue when opening panels
When opening a track group, every panel below was being redrawn.
This caused a lot of slowness on large traces.

In order for mithril to properly map the nodes in an array they
all need keys and the keys must be on the node itself, not the parent.

This change fixes those issues and takes the time to open a track
group on a large trace from ~5s to ~1s.

Bug:124495829
Change-Id: I7fd47ca2dde4150b38c7c41ed1c3e8fdff93e41b