metrics: add proper support for repeated fields
Up until now, we've been making do on repeated fields with strings
representing table names as the repeated field value. However, this is
just unweildy and means we cannot do nested repeated fields properly.

Allow this to happen by introducing a new "RepeatedField" SQL function
which just lumps all the values returned to it and resolves them when
the parent proto attempts to read the value.

Bug: 129747127
Change-Id: I0f2dcbe1b2222e88e5e28588063eeff79d62d787