startup writer: Refactor returning of writers to registry
Chromium was accidentally returning startup writers to the wrong
registry (crbug.com/966307). This patch cleans up the underlying issue
in perfetto, preventing the return of writers to the wrong registry in
the first place. Once landed, we can revert chromium's work-around.

We now proxy the return of a writer via its registry handle and make it
compulsory to return writers created via the registry.

Bug: chromium:966307
Change-Id: I02d6e39d6835719c457c579d46020959bb2db664