trace_processor: add support for extra metrics folder
This CL adds support for external users of shell to register metrics
with shell with more structure than passing a file to --run-metrics
(which is more intended for prototyping and playing with trace
processor).

This has the benefit of allowing fast iteration of builtin metrics as
well by passing the metrics folder path to override all builtin metrics.

Change-Id: I617925f9d795f71818e2325cad9d0d8a0197eaf5