trace_processor: args table: return int-like types as int_value
Currently, the args table never returns bools/uints/pointers. It's not entirely
unreasonable to expect them to be in int_value, which is the approach taken in
this patch. It'll be weird if we ever add a ValueColumn of one of those
specific types, but I'm assuming that we'll remember to update the logic
accordingly.

I didn't test the change besides the newly added unittests.

Change-Id: Id3b8fbdd4105cb319cb80031550367acff1148f6