TraceWriterForTesting: clean up old single-proto assumptions
Before the patch that introduced the TODO, the TraceWriterForTesting
wasn't respecting trace packet boundaries, and was always returning
one merged proto. The android log tests only cared about the contents
of the first packet, that assumption is now more explicit.

Also, changing the single-packet accessor to be more convenient to use:
* simpler return type (the packet proto)
* internal assertion that there's indeed only one packet

Change-Id: Iceee2df3bca1bd4a4fec6c21ce2595ed9a31dc4c