trace_processor: use sched_process_free instead of _exit
This change modifies trace processor to consider sched_process_free
instead of sched_process_exit when doing lifetime tracking of threads.
This is more correct as events can happen in association with a process
after the main thread emits _exit but not after _free.

This change also adds end timestamp tracking for processes as well
defined as the time as the main thread emitting _free.

Change-Id: I253c0c16f719c84b30b72276549993dd0f6f8faa
Bug: 135981504