trace_processor: flatten metadata table to allow RowId indexing
MetadataTable is now based on StorageTable. Note that I've copy-pasted a reduced
version of ArgsTable::ValueColumn for the variadic column. We don't currently
need the full set of types (nor the type casting that we're working on for the
args table), so one instance of copy-paste seemed more pragmatic than making a
generic variadic column impl suitable for both tables (for now).

Note: we discussed the possibility of putting the uniqueness constraint into a
new dedicated tracker (to keep the storage clean of transient state), but tbh
my attempt to implement that ended up involving way too much boilerplate. So
I'm proposing having the "dirty" map in the storage class itself.

Change-Id: I686876178fed151313087bbc4bf8b80e497166fb