trace_processor: support android packages list in proto trace parser
When parsing an android packages list, we turn each package into an arg set,
adding a row to the Metadata table with that arg set id. Parsing & read error
bools are put into the Stats table.

The arg set backfilling is a bit special-casey (it overwrites placeholder rows),
and might have to be changed if we add more use-cases for arg sets to the
metadata table. But this is what I came up with for now. Suggestions welcomed.

Add-on change: proto trace parser unittest string interning now runs real logic
(instead of always returning the default mock return value of 0). And there's
now only a single TraceStorage [nice-]mock in the test.

Change-Id: I4f736f6c6cbbf6f27df0d9fa5ffed4a882554525