CI: Periodically delete stale jobs
If a VM is shut down ungracefully or a docker container
crashes the job would be left in the runqueue. The length
of the runqueue contributes to the autoscaler cost. So
over time crashed jobs left in the queue would cause the
autoscaler to always max-out its capacity.
This CL removes cancelled and too-old jobs form the queue.

Bug: 136726545
Change-Id: If9411df7dc4457af2125b7c3216a935a8cff2d10