shmem: Add an option to drop data instead of stalling on SMB exhaustion
When we run out of free SMB chunks, the writer thread is currently
stalled until the service frees up some new chunks. This works well for
perfetto's traced producers, but is tricky for chromium, because
messages to the TracingService may not be transmitted if IPC threads are
stalled.

This change adds an option to instead fall back to a garbage chunk,
whose data is discarded.

Bug: 136235448
Change-Id: I41f1d117a8c38a5d127c29f572b4a56551bf671c