client api: Add a way to trace directly to a file
This patch adds a TracingSession::SetupWithFile() method that lets the
user provide a file descriptor together with the tracing config. The
tracing service will take ownership of the file and write the trace data
into it.

Note that it's advisable to also set a write flush period (with
TraceConfig::set_flush_period_ms()) when using this API.

Change-Id: I06e8dc11c25a7bd30844e082e29d6556524508a3