shmem: Avoid synchronous commits if in kDrop mode
Moves the logic to send an early commit request synchronously if a lot
of chunks have been buffered to GetNewChunk() and disables it for kDrop
mode.

This is to avoid lock order inversions involving mojo IPC in chromium:
Synchronous flushes would cause a write to the mojo channel, which
acquires the mojo channel's lock, which can happen while we hold some
other lock such as a startup writer's. Mojo code is instrumented to
emit trace events while holding the mojo channel's lock - and emitting
trace events may require acquiring said other lock (startup writer's).

More generally, to avoid this problem in the future, we are planning to
support separating the IPC thread from the tracing service / SMB thread
when using kDrop mode.

Bug: chromium:941318
Change-Id: Ieda74515ba57425621cc46b152057f6954598ae9