client api: Always compile watchdog_posix.cc
The amalgamated build of Perfetto is supposed to be
platform-independent, which means all relevant files need to be compiled
unconditionally for all supported platforms (today: Linux, Mac and
Android), utilizing preprocessor defines to select platform-specific
features.

At the moment, of the targets included in the amalgamated build, only
watchdog_posix.cc is compiled conditinally. This patch fixes this. We'll
land a further presubmit check to avoid any unexpected platform-specific
build files from creeping in.

Bug: 132678367
Change-Id: I44c49f463e04cb21f2ce67bcc58df6b4e13e2c64