Trace processor: Use varint encoding for string size in StringPool
Currently StringPool has a 32K limit on a size of a string. Some trace
fields (e.g. legacy_ftrace_output) can be bigger than that. The limit
is due to the string size being encoded as a 16-bit integer.

This CL introduces support for bigger strings by using VarInt encoding.

Bug: 130786269
Change-Id: Ibda6c2d7eb00a6f10348ffb9b16e0b430503228b