[GPU Profiling] Have a finite set of units.
Previously we copied units from ICU, however, that has too many units that will
never be used and the previous definition of MeasureUnit doesn't have the
flexibility to allow random numerators and denominators. This patch removes the
previous definition and creates a finite set of units that currently we are
certain we will use.

BUG: N/A
Test: build
Change-Id: I5b1fbc3c298032c164c38ce8d94b4df189f96fb9