trace_processor: create sqlite folder and extract initial pieces
To advance the work on the refactor, we need SqliteTable to be isolated in its
own component. Because of this, pull it out along with all its dependenices into
a "sqlite" module.

There are many more classes which could also be pulled out but for now we're doing
the minimum possible.

Bug: 135177627
Change-Id: I8d0d6fe06f1ce1fa850f66718fcba19a6777d52f