Move perfetto's recording controllers at the top
The recording controls that were previously under the "start recording"
entry in the left menu have been moved to the top, so that they are
always visible.
To see the intruction snippet, it is necessary to click the "start
recording" button in the top-right corner.

During a recording, a real progress bar with buffers usage is shown.

Bug: 139716112
Change-Id: I275394cbd2972b06c6b6654a14b4e27ebbbfdfba