trace_processor: add support for strings in db tables
This CL adds support for filtering sorting and querying strings
from the new DB supports as well as allowing the macro tables to
use them by specifying StringId as the column type.

We temporarily create a "common" source set to hold classes which
are needed by the db set to prevent circular dependencies; this
will be cleaned up in future CLs.

Change-Id: I514b0689deacc3ea2c8ec861be71537804731561
Bug: 135177627