tmux script: simplify ssh case, add taskset, remove tee-ing to a log
ssh checks were overriding the is_android case, so you could no longer run
  OUT=out/android_build tools/tmux
and get the "adb shell" case. I also don't think that whether the target is
over an ssh connection has to be tied to cross-compilation. Note that I didn't
have a good setup for testing this, I'm merely assuming that it still works :/.

taskset is used to constrain the cpus the daemons are allowed to run on, useful
when getting a ballpark perf estimate from a device with big/little cores.

tee-ing to a log was a convenience I added earlier, but having the 'tee'
process respond to activity in the tracing daemons can add noise to the traces.

Change-Id: I3f0f83cd1caee42320976dd41e7c1b52ebce83c8