Make trace_processor a host-only target
The trace_processor_shell target is currently very confusing:
it pretends to be a host + device target, but in reality it uses
libprotobuf-full and libprotoc (for dynamic metrics .proto loading)
which is forbidden on-device.
This CL clearly makes trace_processor a host-only target. It also
cleans up some GN files, removing the testonly attribute (and
related hacks in gen_android_bp) and restricting protobuf-full to
host tools only via visibility rules.
It also makes the dependency on libprotoc explicit. It will be
needed by Bazel builds.

Bug: 140126865
Change-Id: Ic8a7827ae83d0ef6c4a607566c732a8798379762