Fix chrome roll: update protoc host toolchain rules
Broken by aosp/1116432, for error see [1].
The problem is that, rightly, chrome doesn't expose libprotoc
and libprotobuf_full on target (!host) toolchains.
We don't really need those targets in the target-toolchain.
However, when running tools/gen_android_bp, the generator doesn't
fully understand toolchains. For the sake of the generator
host-only targets are treated as non-host, and then they are
overridden with |target_host_only|.


[1] https://ci.chromium.org/p/chromium/builders/try/chromeos-amd64-generic-rel/354313

Change-Id: I29dca94b2cb54c7e8acb6e3d3b967760d7686bc1