trace_processor: fix ordering when filtering row vectors
This fixes a subtle bug when filtering a row vector with a predicate. We
would swap the end of the vector with indicies in the middle leading of
elements being out of order.

Instead, work with a read and write pointer and resize to the number of
elements written at the end.

Also add a test which used to fail before.

Bug: 140718238
Change-Id: Ia4fc6ce7faad4890eb20f756ba637db259127e01