trace_processor: make inserting into macro tables more explicit
This CL adds the Row class to all macro generated classes and changes
the Insert method to take a Row.

By adding this layer of indirection, inserting into tables which have a
lot of columns is much easier; now instead of specifing all the fields
(most of which may be null), you can instead specify just the fields you
care about.

We also add a "type" field to all macro tables implicitly - depending
on the Row class used, it will be filled (at the root level) with
the most specific type of Row,

This makes the upcoming track table refactor cleaner as tracks will have
scopes which have a lot of nullable columns

Context: go/perfetto-tp-refactor
Bug: 135177627
Change-Id: I51555f7b0510cba5619d41e4c6a46b8d79de9a5e