Disable watchdog on OSX using compile-time checks
The current logic disables the watchdog OSX using GN-time
check based on the target-os introduced in aosp/1110583.
This broke Android in-tree Mac build. This is because we freeze
the notion of OS once we run the tools/gen_android_bp script,
which then suddenly enables the watchdog on Android Mac host builds.
This fix makes the watchddog enabling logic a compile-time check
which doesn't rely on GN.

Bug: 140765001
Change-Id: Ib91d174b0d05d9c0f921f8642147298a70fa37d0