Enhance open Trace from arrayBuffer peformance
When we opened a trace from an arrayBuffer, we used
'.slice(start,begin)', that creates a copy of the array each time.
Using 'subarray' we save this copy.
the new Uint8array created points to the same 'ArrayBuffer' of the
original UInt8array, but with a different offset and length.

Change-Id: I479a28ec8977333003ba5ec364ca0c74645d6e1d