Fix X86 unwinding.
We were passing the register data in a function expecting a different
order. This happened to work for arm/arm64/x86_64 because the order is
the same in both formats here.

Regs::Read expects user_${arch}_regs, not the vector AsmGetRegs generates,
which is what we provide.

The output of AsmGetRegs corresponds to Regs::RawData.
unwindstack::RegsGetLocal uses AsmGetRegs to write straight into RawData:
https://android.googlesource.com/platform/system/core/+/master/libunwindstack/include/unwindstack/RegsGetLocal.h#91

Test: profile system server on cf_x86. get frames.
Bug: 118428762
Change-Id: I1079d26806fe6badc357285717dc5949423cc36f