perfetto-ui: Persist timespan highlight
The timespan highlight is persisted until the user clicks outside
of the range. This means the user can select individual items inside
the timespan and not lose the highlight.

Also the 'm' key will zoom as well as highlighting the timespan
of the current selection.

See https://goto.google.com/perfetto-multi-selection

Bug:141100137
Change-Id: I5a245e97b30531cec58ad80b622da0e53c41a834