trace_processor: refactor tracks to reflect final design
This CL changes all the track tables and track tracker to reflect the
final agreed shape of these tables.

The main change carried out here is to decouple the idea of the source
of a track from the table it ends up in. Instead, we simply store the
source and the id of the source in the args table - this is useful for
debugging.

Now, all tracks are grouped into track types based not on source but on
what the context of a track is (i.e. thread, process, cpu etc).

However, if a consumer needs to filter tracks based on source, they
should still add a new table and place all their tracks (plus the source)
inside the new subclass - this is done for GPU events.

Context: go/perfetto-tp-refactor
Bug: 135177627
Change-Id: Ie8aa86d84fe9c7266433766d23e8fa247b0966be