trace_processor: change RowMap::Select to return a RowMap
This CL changes RowMap::Select to return a RowMap instead of operating
on *this directly.

We do this to allow for optimizations in cases where we do not need to
copy the entire initial RowMap - think for example the case where we are
filtering on row id and we only have a single row - we can skip copying
full bitvectors and instead just keep the index of the row. This
optimization will be added in a follow-up CL.

Change-Id: I19503db5bb2abe598ef08d346253168d8115962e