test: prefer trace_processor_shell from the host directory.
Before this CL the test script would run inconiditionally
out/xxx/trace_processor_shell. This binary:
On linux/mac builds: was the right binary
On Android and other cross-compiled build: was the host binary
from out/xxx/gcc_like_host that is copied into out/xxx by
the perfetto_host_executable() rule.

The issue is that the logic of perfetto_host_executable()
is not correct in the case of non-hermetic builds, where the
generated executable depends on other libraries (e.g. the
self-built libc++.so for asan/lsan).
This is okay for humans, because the only use case developers
care about, for host+target builds, is typing
out/xxx/trace_processor_shell in android builds.
Instead, in tests, use the explicit out/xxx/gcc_like_host
directory, so the binary uses the right version of libc++.so.

Bug: 141319350
Change-Id: Id69d96a01d144ae91117e98fa8f80f2607ac09d3