trace_processor: fix dcheck failure when we get an out of order upid counter
This CL fixes the issue where we got an out of order upid event where we
needed to resolve utid -> upid. In that situation, we would try and
resolve the upid even though the row_id returned from PushCounter was
invalid.

Fix the issue by not creating a pending entry when we get a invalid row
id.

Bug: 141267734
Bug: 142713794
Change-Id: Iaf44b7e00f3903c8017ec3616a936fcf164ca655