Remove the call to std::partition, and replace with a readable for loop.
The call to partition was a small optimization to have valid consumers
near the front of the vector so that FindConsumer hit matches faster,
however this micro optimization isn't worth the hit in readability so
just store the consumer and check. Also replace std::distance with

std::count_if to more clearly express the DCHECK we're enforcing.
Change-Id: I84b2bbc113f5574cccc39dcb78b780bf08af0fd2