Add TCP/IP support to UnixSocket
Adds the ability to create TCP sockets. This
will be used by the trace processor HTTP interface
(See bug).
Arguably the class should be renamed to just Socket,
but that is a larger change and perhaps not worth given
the nicher use case for TCP.
This CL also gets rid of the default args and makes all
call-sites explicit to avoid bugs in future.

Bug: 143074239
Change-Id: Ib3a66e1f33ff2c025defb0ecd066b0e0e491e2e9