processor: Expose a C++ API for ExportJson to chrome
Turns out this makes a few things simpler for chrome:
(1) Chrome uses argument/metadata/label filters, which are predicates.
(2) Makes it possible to pass in something other than an FD for output.

(1) is tricky to solve declaratively given the complexity of these
filters and the existing legacy code in chromium. (2) could be solved
by providing a memory-backed FILE* on posix (fmemopen) - but not so
easily on windows.

Thus this new API allows providing the filter predicates (impl TODO)
and a custom writer for output.

Bug: 130786269
Change-Id: I002e67d5160157bd494c85dbe2c159695bd24c25