trace_processor: add range mode to RowMap
This CL adds a "range" mode to Row which allows a block of rows to be
stored efficiently without the memory/cpu overhead of BitVector.

Specifically, there are two instances where this is valuable:
1. When we create the initial RowMap for filtering, we no longer need to
   actually allocate memory for the bitvector but instead just store a
   range with all the rows in the table.
2. When we have nested subqueries on id columns, we end up with only
   having zero or one rows in a RowMap and it's inefficient to create a
   huge bitvector to store this one value.

A follow up CL will actually start making use of this feature with
benchmarks to measure performance.

Context: go/perfetto-tp-refactor
Bug: 135177627
Change-Id: I3783eaf2d9aab2d58ab4d6ca5c9c6b2979cf3c76