UI; unify File/ArrayBuffer/fetch() into TraceStream
This CL unifies the three code paths for loading traces
introducing a TraceStream class. This class exposes
a unified interface for loading a trace in chunks and
provides three implementations for opening a trace from:
(1) A File object; (2) An ArrayBuffer (for window.open());
(3) A url.
Also this CL passes WASM input on a dedicated buffer
rather than on the WASM stack. This allows the bump the
chunk size when loading from file from 1MB to 32MB, upping
significantly the trace load speed because of hitting
JS <> WASM rundtrips less frequently.

Bug: 143074239
Change-Id: I97660b47d26378428be514fc79667cf9408dbdfe