protozero: Make generated cpp enum/class order stable
Updates protozero's CppObjGenerator to use std::vectors to keep track
of enums and classes to generate, so that their order in the generated
files remain stable.

Previously, the std::sets storing pointer addresses would result in
indeterministic ordering due to changing addresses - this was causing
deterministic build failures in chromium.

Bug: chromium:1017699
Change-Id: I712617612a269662ac6a0d08efb1c84050e4d071