compact sched_waking: event decoding in trace_processor
Compact waking events are now handled by the sched_event_tracker, which populates the
raw & instants tables. Handling of the default-encoded events hasn't changed.

The emitting pid is inferred from the [pending] scheduling slices, in other
words, sched_switch events. So if you have compact sched_waking, but no
sched_switch, we skip all waking events during parsing. I think this is fine,
and just needs some extra documentation. We could instead start encoding the
emitting pid in the trace to remove this limitation, but I'm not sure who'd want
*compact* sched_waking without any sched_switch events.

Sorter: I've kept the approach of having all events in the same per-cpu queue.
We can theoretically do better with multiple queues, but that also requires
further changes to the sorter that are outside the scope of this cl.  In
practice, this approach does not give significant overheads on typical traces
(especially if there's no write_to_file, as then we sort each queue at most once
at the end).

Change-Id: I1deda1d433b696dd16a5f127604c9ab8f6f8cdc9