trace_processor: fix broken iteration logic in thread and proces tables
This CL fixes the broken logic in iteraing through the threads and
processes table when "ORDER BY utid/upid DESC" caluse was present in a
query.

Before this CL, we would have tried to iterate to uint32::max row
because we would have underflowed the index when iterating backwards.

Instead, change both tables to always count the offset from min up to
max and also change max to a exclusive index to make it more
standardised.

Change-Id: I6e81d43ef8a53970ca0c7ed70d9a759279f81800