trace_processor: rework best index logic again
In a previous CL we added a bunch of boilerplate around removing and
sorting QueryConstraint sets. Unfortunately, this was still not
powerful enough and was leading to some really funky code.

Just expose the raw query constraints for subclasses to modify as
a prepatory step instead of trying to insulate them from errors.

Change-Id: Ifca519475c48c84194cf8b3cebf19da0ec0777b0