ui: Fix persisting of target selection on record page
Refactores the way we store the selected target device / os in the
state to fix issues where the target drop down didn't show the selected
target OS/device correctly.

Now, the target is stored separately from the config in the state no
matter whether an OS or device is selected.

Bug: 144072582
Change-Id: I199bc2daa28b85f88ac0dc1fb8932c737adc4a08